code/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
psutil==5.9.8
pydantic>=2.7.2
pydantic-settings>=2.2.1
websockets==12.0
python-multipart==0.0.9
typing-extensions>=4.12.0
langchain==0.1.12
langchain-openai==0.0.8
openai==1.12.0
beautifulsoup4==4.12.3
SQLAlchemy==2.0.40
aiosqlite==0.19.0 